Implement initial version of BanyanHDF5.jl
