+++ BanyanArrays/src/BanyanArrays.jl
@@ -1,46 +1,26 @@
 module BanyanArrays
 
 using Banyan,
-    AWS,
-    AWSCore,
-    AWSS3,
-    Banyan,
-    Downloads,
-    FileIO,
-    FilePathsBase,
-    HDF5,
     MPI,
-    Missings,
-    Random,
     Serialization
 
 export Array, Vector, Matrix
 export read_hdf5, write_hdf5
 export ndims, size, length, eltype
-export fill, zeros, ones, trues, falses
-export map, mapslices, reduce, sort, sortlices
+export fill, zeros, ones, trues, falses, collect
+export map, mapslices, reduce, sort, sortlices, getindex
 
 export ReadBlockJuliaArray,
     ReadGroupJuliaArray,
     WriteJuliaArray,
     CopyFromJuliaArray,
     CopyToJuliaArray,
-    ReadBlockHDF5,
-    ReadGroupHDF5,
-    WriteHDF5,
-    CopyFromHDF5,
-    CopyToHDF5,
     SplitBlock,
     SplitGroup,
     Rebalance,
     Consolidate,
-    Shuffle,
-    collect,
-    getindex
+    Shuffle
 
-export RemoteTableSource, RemoteTableDestination
-
-include("locations.jl")
 include("array.jl")
 include("utils_pfs.jl")
 include("pfs.jl")
