+++ BanyanHDF5/test/Project.toml
@@ -1,9 +1,11 @@
 [deps]
 AWSS3 = "1c724243-ef5b-51ab-93f4-b0a88ac62a95"
 Banyan = "706d138b-e922-45b9-a636-baf8ae0d5317"
+BanyanArrays = "369465de-032e-4609-9dcf-82b89c370a7b"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 FilePathsBase = "48062228-2e41-5def-b9a4-89aafe57970f"
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 ReTest = "e0db7c4e-2690-44b9-bad6-7687da720f89"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
@@ -11,5 +13,6 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 Banyan = "0.3.1"
+BanyanArrays = "0.2.1"
 ReTest = "0.3.2"
 julia = "1.6"
